Change README theme instructions
If you do a git clone as the instructions point out, the directory should be named in themes as `anpu-zola-theme` not `anpu`. Otherwise zola will complain with `No `theme.toml` file found. Is the `theme` defined in your `config.toml` present in the `themes` directory and does it have a `theme.toml` inside?`